+++ src/main/java/net/soulsweaponry/items/crossbow/KrakenSlayerCrossbow.java
@@ -1,32 +1,28 @@
 package net.soulsweaponry.items.crossbow;
 
-import net.minecraft.enchantment.Enchantments;
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.projectile.PersistentProjectileEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.recipe.Ingredient;
-import net.minecraft.world.World;
 import net.soulsweaponry.config.ConfigConstructor;
-import net.soulsweaponry.items.bow.KrakenSlayer;
-import net.soulsweaponry.util.TooltipAbilities;
-import net.soulsweaponry.util.WeaponUtil;
+import net.soulsweaponry.items.abilities.customarrows.ThirdShotTrue;
 
 import java.util.function.Supplier;
 
 public class KrakenSlayerCrossbow extends ModdedCrossbow {
 
+    private static final ThirdShotTrue THIRD_SHOT_CROSSBOW = new ThirdShotTrue(
+            ConfigConstructor.kraken_slayer_crossbow_true_damage,
+            ConfigConstructor.kraken_slayer_crossbow_bonus_true_damage_per_level,
+            "CROSSBOW",
+            (int) ConfigConstructor.kraken_slayer_crossbow_stacks_per_shot,
+            ConfigConstructor.kraken_slayer_crossbow_bonus_stacks_per_shot_per_level,
+            (int) ConfigConstructor.kraken_slayer_crossbow_max_stacks_until_true_damage
+    );
+
     public KrakenSlayerCrossbow(Settings settings, Supplier<Ingredient> repairIngredientSupplier) {
         super(settings, createConfig((int) ConfigConstructor.kraken_slayer_crossbow_pull_time_ticks,
                 ConfigConstructor.kraken_slayer_crossbow_damage, ConfigConstructor.kraken_slayer_crossbow_bonus_velocity),
                 repairIngredientSupplier);
-        //this.addTooltipAbility(TooltipAbilities.FAST_PULL, TooltipAbilities.THIRD_SHOT);
-    }
-
-    public PersistentProjectileEntity getModifiedProjectile(World world, ItemStack bowStack, ItemStack arrowStack, LivingEntity shooter, PersistentProjectileEntity originalArrow) {
-        float bonus =  WeaponUtil.getLevel(bowStack, Enchantments.QUICK_CHARGE) / 4f;
-        PersistentProjectileEntity newProjectile = KrakenSlayer.getKrakenSlayerProjectile(world, bowStack, shooter, originalArrow.getDamage() + bonus, ConfigConstructor.kraken_slayer_bonus_true_damage);
-        originalArrow.setDamage(originalArrow.getDamage() + bonus);
-        return newProjectile;
+        this.addAbility(THIRD_SHOT_CROSSBOW);
     }
 
     @Override